backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
supabase
httpx
pydantic
fastapi
uvicorn
pydantic
joblib
lightgbm
scikit-learn
pandas
numpy
